refactor: 미래시간 검증을 도메인에서 하도록 변경
