+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -2,10 +2,12 @@ package roomescape.reservation.domain;
 
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.time.LocalTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NonNull;
+import roomescape.error.ReservationException;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
 
@@ -29,6 +31,7 @@ public class Reservation {
     public Reservation(@NonNull final String name, @NonNull final LocalDate date,
                        @NonNull final ReservationTime reservationTime,
                        @NonNull final Theme theme) {
+        validateFutureOrPresent(date, reservationTime.getStartAt());
         this.id = null;
         this.name = name;
         this.date = date;
@@ -36,7 +39,11 @@ public class Reservation {
         this.theme = theme;
     }
 
-    public LocalTime extractTime() {
-        return time.getStartAt();
+    private void validateFutureOrPresent(final LocalDate date, final LocalTime time) {
+        final LocalDateTime reservationDateTime = LocalDateTime.of(date, time);
+        final LocalDateTime currentDateTime = LocalDateTime.now();
+        if (reservationDateTime.isBefore(currentDateTime)) {
+            throw new ReservationException("예약은 현재 시간 이후로 가능합니다.");
+        }
     }
 }
